requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-allauth==0.51.0
django-bootstrap3==22.1
django-bootstrap5==22.1
django-braces==1.15.0
django-cors-headers==3.13.0
django-extra-views==0.14.0
django-filter==22.1
django-pg-utils==0.1.5
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==14.0.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
misaka==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.1
Pillow==9.2.0
psycopg2 @ file:///C:/ci_310/psycopg2_1642085998310/work
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
types-cryptography==3.3.23
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
wincertstore==0.2
